astgen: install resources/ to share/cppmm/resources/
Relocate the astgen header resource files under share/cppmm/resources/
for consistency with the packaging standards reflected in
GNUInstallDirs.
A cmake cache variable is provided to allow for this location to be
configured by users / distributors.

Add a #define provided by the build system so that astgen.cpp can
be told where to find its resources rather than hard-coding a specific
location.