Add CPPMM_MANUAL attribute
This simply forces the binding function signiature to be included in the
AST. This is very useful e.g. for when the target library has a default
argument that is a PITA to bind and only shows up there, like the use of
SFINAE in USD, or the optional allocator arguments to std::string.