asttoc: skip linking libraries when building docs
docs.rs does not need to link our crates. It only needs the
FFI metadata. We can skip linking when building for docs.rs.

This removes the need to have our dependent libraries available and
makes it possible for docs.rs to build have documentation for both
the sys and high-level rs crates.

https://docs.rs/about/builds#detecting-docsrs
https://gist.github.com/virtualritz/2fe2d61ee2bae6d211acbef55aac41fe