cppmm: add support for building documentation under docs.rs
Update the build.rs process so that it generates an additional
cppmmabi_generated.rs that is intended to be committed into
the repository and deployed to docs.rs.

The intended workflow is that the generated file should be
refreshed whenever a new version is being published.

This allows docs.rs to generate documentation without having
to have cmake or any of the library dependencies available.

docs.rs is building for x86_64 Linux only, so as long as we
upload compatible cppmmabi_generated.rs files then we can
benefit from having our documentation hosted on docs.rs.